Streamline logic for selecting login flow

+++ strengthjournal-app/src/app/header/header.component.ts
@@ -1,6 +1,7 @@
 import { Component, Inject, Input, OnInit } from '@angular/core';
 import { AuthService } from '@auth0/auth0-angular';
 import { DOCUMENT } from '@angular/common';
+import { environment } from 'src/environments/environment';
 
 @Component({
   selector: 'app-header',
@@ -12,8 +13,13 @@ export class HeaderComponent implements OnInit {
   @Input() userFullName: string = '';
 
   logout() {
-    localStorage.clear();
-    this.auth.logout({ returnTo: document.location.origin });
+    if (environment.useResourceOwnerFlow) {
+      localStorage.clear();
+      window.location.replace('/');
+    }
+    else {
+      this.auth.logout({ returnTo: document.location.origin });
+    }
   }
 
   constructor(@Inject(DOCUMENT) public document: Document, public auth: AuthService) { }
